GET punto residuo by ID

+++ src/main/java/org/circle8/controller/response/ResiduoResponse.java
@@ -10,17 +10,17 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Builder(toBuilder = true)
 public class ResiduoResponse implements ApiResponse {
-	public int id;
+	public long id;
 	public LocalDateTime fechaRetiro;
 	public LocalDateTime fechaCreacion;
 	public String puntoResiduoUri;
-	public Integer puntoResiduoId;
+	public Long puntoResiduoId;
 	public PuntoResiduoResponse puntoResiduo;
 	public TipoResiduoResponse tipoResiduo;
 	public String recorridoUri;
-	public Integer recorridoId;
+	public Long recorridoId;
 	public RecorridoResponse recorrido;
 	public String transaccionUri;
-	public Integer transaccionId;
+	public Long transaccionId;
 	public TransaccionResponse transaccion;
 }
